Try to override Jackson to 2.8.9
For using some library in Toree we need newer Jackson than 2.6.5 which is in Spark 2.0-2.2.
In this commit I try to override version of Jackson, including some modules.
It seems that in the Toree fat JAR there a correct version, but still in
Jupyter notebook with Toree Scala kernel it loads Jackson 2.6.5 from somewhere :(

new com.fasterxml.jackson.core.json.PackageVersion().version()